Player positioning api, move play sound to stable.
The play sound packet was not tested correctly. It works perfectly.
The tests I performed before were when the player connects, but
didnt knew that bungee does not handle client directed packets
on connect. The ProxiedPlayer#playSound was adjusted to work with
the positioning api and i removed some accidentaly added imports
in the remove entity rewrite patch.