2 new patches
- Fill missing manifest information (by Janmm14)
This is not that "interesting". It's all java and information related things

- Rewrite LogEvents to contain the source jars in stack traces
Ported over from Paper with small modifications so it works with java 8.
These small modifications are that we don't use classloader name (which is
java >= 9+ only feature), rather we try to find the pluginclassloader which
loaded the exception's source class and obtain the source jar from there.