Reset old Title from other server on Upstream initializing
Some servers are sending titles that do not disappear.
Example: In hypixel, if you're marked as afk a title appears and
if you disconnect and join on another server and you have not moved again
on hypixel at the time you disconnected, you will have the same title on
the other server and it won't disappear.